ref: remove old Player class, rename playerRefactor to Player
